requirements.txt
Flask==2.3.3
python-docx==1.1.0

